feat(feature_component_table): made the entity service, repository, etc... in hexagonal for the seller table so it is more prepared for the component table #5

+++ ApiMasterPCBuilder/src/main/java/es/iespuertodelacruz/ohjm/apimasterpcbuilder/infrastructure/adapter/secundary/repository/IUserEntityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface IUserEntityRepository extends JpaRepository<UserEntity, Integer> {
+public interface IUserEntityRepository extends JpaRepository<UserEntity, Long> {
     @Query(value = "SELECT * FROM USERS u WHERE u.nick = :nick", nativeQuery = true)
     UserEntity findByNick(@Param("nick") String nick);
